+++ crates/cdk-http-client/Cargo.toml
@@ -14,6 +14,7 @@ readme = "README.md"
 default = ["bitreq"]
 bitreq = ["dep:bitreq"]
 reqwest = ["dep:reqwest"]
+bip353 = ["dep:hickory-resolver", "dep:rustls", "dep:ring"]
 tor = [
     "dep:arti-client",
     "dep:arti-hyper",
@@ -39,12 +40,15 @@ tracing.workspace = true
 url.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-bitreq = { version = "0.3.4", features = ["async", "async-https-rustls", "json-using-serde", "proxy"], optional = true }
+bitreq = { version = "0.3.4", features = ["async", "async-https-rustls", "json-using-serde", "proxy"], default-features = false, optional = true }
+hickory-resolver = { version = "0.25.2", optional = true, features = ["dnssec-ring"] }
+ring = { version = "0.17", optional = true }
+rustls = { workspace = true, optional = true }
 arti-client = { version = "0.19.0", optional = true, default-features = false, features = ["tokio", "rustls"] }
 arti-hyper = { version = "0.19.0", optional = true }
 getrandom = { version = "0.2", optional = true }
 http = { version = "0.2", optional = true }
-hyper = { version = "1.9.0", optional = true, features = ["client", "http1", "http2"] }
+hyper = { version = "0.14", optional = true, features = ["client", "http1", "http2"] }
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"], optional = true }
 tls-api = { version = "0.9", optional = true }
 tls-api-native-tls = { version = "0.9", optional = true }
