+++ crates/cdk/Cargo.toml
@@ -16,7 +16,7 @@ wallet = ["dep:futures", "cdk-common/wallet", "cdk-common/http", "dep:rustls"]
 nostr = ["wallet", "dep:nostr-sdk", "cdk-common/nostr"]
 npubcash = ["wallet", "nostr", "dep:cdk-npubcash"]
 mint = ["dep:futures", "cdk-common/mint", "cdk-common/http", "cdk-signatory"]
-bip353 = ["dep:hickory-resolver", "cdk-common/bip353"]
+bip353 = ["cdk-common/bip353", "cdk-http-client/bip353"]
 bench = []
 http_subscription = []
 tor = [
@@ -57,7 +57,6 @@ tokio-util.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 cdk-http-client.workspace = true
-hickory-resolver = { version = "0.25.2", optional = true, features = ["dnssec-ring"] }
 tokio = { workspace = true, features = [
     "rt-multi-thread",
     "time",
